feat: change datatype to accomodating error

+++ include/models.h
@@ -7,8 +7,8 @@ struct SensorData
     float temperature;
     float humidity;
     float nh3;
-    uint16_t lux;
-    uint16_t windSpeed;
+    int lux;
+    int windSpeed;
 
     String toString()
     {
